Fix #230 - lookup breaking
The problem is that
https://7dtd.illy.bz/browser/binary-improvements/MapRendering/Web/API/GetPlayerList.cs#L28
returns 25 players by default. The code assumes the user will be in the
list of players. Sadly the machinepack library doesn't let you specify
more anyways. Plus I don't think it actually matters since its really
just grabbing a list of players, and it'll return the most recent 25
users.

So if someone is looking up an older user, we already know we have the
db version and its unlikely changed. So return the db version when it
cant find the return from the server.